+++ test/amp/MergeTest/MergeTest.cpp
@@ -340,7 +340,7 @@ TEST(sanity_merge_bolt_vect, wo_control_ints)
 	std::vector<int>   stdmerge(Size);
 
 	for (size_t i = 0; i < aSize; i++) {
-		B[i] = A[i] = i;
+		B[i] = A[i] = (int)i;
 	}
 
 	std::sort(A.begin(), A.end());
