+++ test/cl/GatherTest/GatherTest.cpp
@@ -3492,9 +3492,9 @@ TEST( HostMemory_int, OffsetGatherIfPredicateMedium )
     std::vector<int> stencil (myStdVectSize,0);
 	for (size_t i = 0; i < distance; i++)
     {
-        map[i] = s_offset + i;
-        input[i] =  i + 2 * i;
-        stencil[i] =  i + 5 * i;
+        map[i] =(int) (s_offset + i);
+        input[i] =  (int)(i + 2 * i);
+        stencil[i] =  (int)(i + 5 * i);
     }
     std::random_shuffle( map.begin(), map.end() );   
     is_even iepred;
@@ -3542,8 +3542,8 @@ TEST( HostMemory_int, OffsetGatherPredicateMedium )
     std::vector<int> map (myStdVectSize,0);	
 	for (size_t i = 0; i < distance; i++)
     {
-        map[i] = s_offset + i;
-        input[i] =  i + 2 * i;
+        map[i] = (int)(s_offset + i);
+        input[i] = (int)( i + 2 * i);
     }
     std::random_shuffle( map.begin(), map.end() );
